Removed Spring Batch for the Time Being
Still not sure about the correct way to configure Spring Batch. The code was starting to get messy in my attempts to make it work, so the related code has been replaced temporarily with a single method call that works in a similar amount of time. Duplicate user validation needs to be moved to using a staging table as the DB will be faster at operations on sets. That should pull the processing time to about 40 seconds for 100k entries or 5 seconds for 10k entries with or without Spring batch