Major Refactor of REST Configuration with Swagger Codegen
In order to maintain calls between services, all REST endpoints have been shifted to OpenAPIs codegen server stubs, with a separate Maven module for generating the client libraries. Both the server-stubs and the client-libraries are built from the same YAML spec, and with proper CICD practices should be properly versioned. This allows other services to simply include the client library as a dependency to gain access to it's REST endpoints and DTOs that it will be communicating with. These DTOs are also what is transmitted over Kafka topics, enabling seamless communication there as well.

* Separated into two distinct modules, server and client
* Implemented generated server stubs
* Removed references to common library DTOs in favor of generated ones